requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
django-ckeditor
django-cleanup
django-crispy-forms
django-rest-swagger
djangorestframework-simplejwt
drf-yasg
celery
redis